build(release): Publish this project to JSR
Users can run via the JSR published script.
