+++ deno.json
@@ -1,22 +1,24 @@
 {
-  "fmt": { "exclude": ["LICENSE", ".github/**/*.md"] },
+  "name": "@5ouma/reproxy",
+  "version": "1.0.0",
+  "exports": { ".": "./src/app.ts", "./serve": "./src/server.ts" },
+  "publish": {
+    "include": ["LICENSE", "README.md", "deno.json", "src/"],
+    "exclude": ["**/*.test.ts", "src/libs/test_utils.ts"]
+  },
+  "fmt": { "exclude": ["LICENSE", "README.md", ".github/**/*.md"] },
   "tasks": {
     "run": "deno run --env='.env' --allow-env='REPOSITORY_OWNER,REPOSITORY_NAME,REPOSITORY_PATH,GITHUB_TOKEN' --allow-net='0.0.0.0,api.github.com'",
-    "start": "deno task run ./src/server.ts",
-    "dev": "deno task run --watch ./src/server.ts",
+    "start": "deno task run src/server.ts",
+    "dev": "deno task run --watch src/server.ts",
     "test": "deno test --allow-env='REPOSITORY_OWNER,REPOSITORY_NAME,REPOSITORY_PATH,GITHUB_TOKEN' --allow-net='0.0.0.0,api.github.com' --parallel --shuffle",
     "cov": "deno task test --coverage && deno coverage --lcov > coverage.lcov"
   },
   "imports": {
     "@hono/hono": "jsr:@hono/hono@4.5.8",
-    "@octokit/rest": "https://esm.sh/@octokit/rest@21.0.1",
     "@std/assert": "jsr:@std/assert@1.0.2",
     "@std/http": "jsr:@std/http@1.0.2",
-    "@std/path": "jsr:@std/path@1.0.2"
-  },
-  "deploy": {
-    "project": "reproxy",
-    "include": ["./deno.json", "./src/"],
-    "entrypoint": "./src/server.ts"
+    "@std/path": "jsr:@std/path@1.0.2",
+    "@octokit/rest": "npm:@octokit/rest@21.0.1"
   }
 }
