+++ src/app.test.ts
@@ -1,7 +1,7 @@
 import { assertEquals } from "@std/assert";
 import { STATUS_CODE } from "@std/http/status";
 
-import { app } from "./server.ts";
+import { app } from "./app.ts";
 import {
   exportRepo,
   testRef,
